% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepm.R
\name{forecast_comparison}
\alias{forecast_comparison}
\title{A function to compare forecast performance of nepm to lm}
\usage{
forecast_comparison(
  pdat,
  x_names,
  y,
  id,
  time,
  max_time_out = 0,
  start_time = NULL
)
}
\arguments{
\item{pdat}{The panel dataset as a dataframe.}

\item{x_names}{Character vector giving the names of the covariates. Should be column names in pdat.}

\item{y}{Character name of the dependent variable in pdat.}

\item{id}{Character name of the unit/node id in pdat. Should be a character variable.}

\item{time}{Character name of the numeric time variable in pdat.}

\item{max_time_out}{Integer, number of time periods a node can be out of the data in the beginning.}

\item{boot}{Indicator of whether to use bootstrapping to calculate uncertainty measures.}

\item{nboot}{Integer, number of bootstrap iterations.}

\item{start_tim}{Integer, the time period at which to start the forecasting experiment. Defaults to half the time points.}
}
\value{
list with a character vector of edges inferred, a dataframe that can be used to run nepm, and a formula that combines the edges and the covariates in the model.
}
\description{
A function to compare forecast performance of nepm to lm
}
