% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepm.R
\name{nepm}
\alias{nepm}
\title{A function to run NEPM}
\usage{
nepm(pdat, x_names, y, id, time, max_time_out = 0, test_edges = F, nperm = 100)
}
\arguments{
\item{pdat}{The panel dataset as a dataframe.}

\item{x_names}{Character vector giving the names of the covariates. Should be column names in pdat.}

\item{y}{Character name of the dependent variable in pdat.}

\item{id}{Character name of the unit/node id in pdat. Should be a character variable.}

\item{time}{Character name of the numeric time variable in pdat.}

\item{max_time_out}{Integer, number of time periods a node can be out of the data in the beginning.}

\item{test_edges}{Logical, indicator of whether to test the significance of the edges (as a block)}

\item{nperm}{integer, number of permutations to use in the edge permutation test}
}
\value{
list with a character vector of edges inferred, a dataframe that can be used to run nepm, and a formula that combines the edges and the covariates in the model.
}
\description{
A function to run NEPM
}
\examples{
library(nepm)

## generative process for NEPM
# set some parameters
times <- 50 # time periods in data
nodes <- 50 # number of nodes
nties <- 20 # number of ties

# simulate some data
set.seed(1001416)
node_id <- paste("n",1:nodes,sep="")
possible_ties <- rbind(t(combn(node_id,2)), t(combn(node_id,2))[,c(2,1)])
ties <- possible_ties[sample(1:nrow(possible_ties),nties),] # select the ties

# make model parameter values
gamma <- .45 # set parameter to force stationarity
beta <- c(.5,-.5)
sig <- 1

# generate a covariate that is not time-varying (easy to make time varying)
x <- rnorm(nodes)

# generate new time period given previous time point
sim_yt <- function(ytm1,x,beta,gamma,ties,sig){
  # ytm1 is a nodes x 1 vector
  # x is a nodes x k matrix where k is the number of covariates
  # beta is a (k + 1) x 1 vector of regression coefficients
  # gamma is a nties x 1 vector of tie effects
  # ties is a nties x 2 matrix where the columns are in order of sender/receiver
  # sigsq is the standard deviation of the error term

  nodes <- length(ytm1)

  # make adjacency matrix of network effects
  amat <- matrix(0,nodes,nodes)
  rownames(amat) <- paste("n",1:nodes,sep="")
  colnames(amat) <- paste("n",1:nodes,sep="")
  amat[ties] <- gamma

  amat <- t(amat)

  # network effect
  net_effs <- amat\%*\%cbind(ytm1)

  # new y
  cbind(1,x)\%*\%cbind(beta) + net_effs + rnorm(nodes,sd=sig)

}


# matrix of simulated time periods
ytm1 <- rnorm(nodes,sd=sig)
panel_data <- NULL
for(t in 1:(1000+times)){
  yt <- sim_yt(ytm1,x,beta,gamma,ties,1)
  panel_data <- rbind(panel_data,t(yt))
  ytm1 <- yt
}

# take last time steps as final data
panel_data <- panel_data[1001:(1000+times),]

# long data
long_data <- NULL
for(i in 1:ncol(panel_data)){
  dati <- data.frame(id = rep(paste("n",i,sep=""),nrow(panel_data)),y=panel_data[,i],
                     time=1:nrow(panel_data),x=x[i])
  long_data <- rbind(long_data,dati)

}

system.time(nepm_test <- nepm(long_data,x_names="x",
                              y= "y",
                              id ="id",
                              time = "time"))

nepm_estimate <- lm(nepm_test$nepm_formula,data=nepm_test$new_pdat)
}
